+++ CHANGELOG.md
@@ -14,6 +14,7 @@ In Development
   after parsing the manifest
 - The default `--jobs` value now equals the number of available CPU cores or
   20, whichever is lower
+- Add `--ok-errors` option
 
 v0.1.0-alpha.2 (2025-01-06)
 ---------------------------